==> 8_Taxonomy_comparison.Rmd <==
---
title: "CompareTaxonomy"
author: "David Garcia"
date: '2022-11-29'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Load libraries
```{r, warning=FALSE}
library(taxize)
library(tidyverse)
library(data.table)
library(readr)
library(tools)
```

# Create output folder
```{r}
dir.create(path = "./tables/enrichment_families")
```

# Retrieve taxonomy for GMRepo taxa using taxize.
This step can take a while because taxize requests the taxonomy to the ncbi and the taxa list is long.
If you don't want to repeat this step, you can use the output file
"./supporting_metadata/GMrepo_taxonomy.csv"

```{r}
gutMicrobiome <- read_tsv("./supporting_metadata/GMREPO_relative_abundance_of_all_species_genus_in_all_phenotypes_summary.tsv",
                          skip = 2)

colnames(gutMicrobiome)[3] <- "id"

gutMicrobiome$id <- as.character(gutMicrobiome$id)

taxonID <- gutMicrobiome %>% distinct(id)

taxonomy_results <- list()

error_ids <- list()

max_retries <- 3

for (i in taxonID$id) {
  retry_count <- 0

  while(retry_count < max_retries) {
    retry_count <- retry_count + 1
    tryCatch({
      taxonomy <- classification(i, db = "ncbi")
      taxonomy_results[[as.character(i)]] <- taxonomy
      break
    },
    error = function(e) {
      cat("Error for ID", i, ":", conditionMessage(e), "Attempt", retry_count)
      error_ids <- c(error_ids, id)
    })
  }
}

save(taxonomy_results, file = "./RData/taxonomy_results_list.RData")
```

# Format GMRepo taxonomy table

```{r}
load("./RData/taxonomy_results_list.RData")

taxonomy <- data.frame()

# i <- "1678"

for (i in names(taxonomy_results)) {
  tempTaxonomy <- taxonomy_results[[i]]
  tempTaxonomy <- rbindlist(tempTaxonomy)
  tempFamily <- tempTaxonomy %>% filter(rank == "family")
  tempPhylum <- tempTaxonomy %>% filter(rank == "phylum")
  taxonomy <- rbind(taxonomy, cbind(i, tempFamily, tempPhylum))
}

# Exclude taxa with unknown family classification

idx <- which(is.na(taxonomy$name) == TRUE)

taxonomy <- taxonomy[-idx, ]

# Change headings

colnames(taxonomy) <- c("NCBI_taxon_id", "family", "rank_1",
                        "NCBI_family_id", "phylum", "rank_2",
                        "NCBI_phylum_id")

write_csv(taxonomy, "./supporting_metadata/GMrepo_taxonomy.csv")
```

# Load samples taxonomy data
```{r}
load("./RData/ps_merge.Rda")

taxonomy <- data.frame(ps_merge@tax_table)

taxonomy <- rownames_to_column(taxonomy, var = "ASV_id")
```

# ARCOL_Leclerc_et_al_2021

```{r}
asv_list_files <- list.files("./tables/enrichment_analysis/ARCOL_Leclerc_et_al_2021/", full.names = TRUE)

asv_sample_id <- file_path_sans_ext(list.files("./tables/enrichment_analysis/ARCOL_Leclerc_et_al_2021/"))

csv <- lapply(asv_list_files, read.csv)

project_table <- as.data.frame(csv[[1]])

for (i in 2:length(csv))
{
  project_table <- full_join(project_table, as.data.frame(csv[[i]]))
}

colnames(project_table) <- c("ASV_id", asv_sample_id)

# convert to presence/absence data.frame (i.e. >)

project_table[is.na(project_table)] <- 0


project_table[, 2:length(project_table)] <-
  ifelse(project_table[, 2:length(project_table)] > 1, 1, 0)

test <- project_table

test <- test[, -c(2, 5)]

idx <- data.frame(which(test != 0, arr.ind = TRUE))

row_id <- data.frame(table(idx$row))

row_id <- row_id %>% filter(Freq > 2)

row_id$Var1 <- as.numeric(as.character(row_id$Var1))

important <- test[row_id$Var1, ]

important <- merge(important, taxonomy, by = "ASV_id")

important %>%
  group_by(Species) %>%
  summarise_all(na.omit)

write.csv(important, "./tables/enrichment_families/ARCOL_Leclerc_et_al_2021_enriched.csv")
```

# M-ARCOL_Deschamps_et_al_2020

```{r}
asv_list_files <- list.files("./tables/enrichment_analysis/M-ARCOL_Deschamps_et_al_2020/", full.names = TRUE)

asv_sample_id <- file_path_sans_ext(list.files("./tables/enrichment_analysis/M-ARCOL_Deschamps_et_al_2020/"))

csv <- lapply(asv_list_files, read.csv)

project_table <- as.data.frame(csv[[1]])

for (i in 2:length(csv))
{
  project_table <- full_join(project_table, as.data.frame(csv[[i]]))
}

colnames(project_table) <- c("ASV_id", asv_sample_id)

# convert to presence/absence data.frame (i.e. >)

project_table[is.na(project_table)] <- 0


project_table[, 2:length(project_table)] <-
  ifelse(project_table[, 2:length(project_table)] > 1, 1, 0)

test <- project_table

test <- test[, -c(2, 5)]

idx <- data.frame(which(test != 0, arr.ind = TRUE))

row_id <- data.frame(table(idx$row))

row_id <- row_id %>% filter(Freq > 2)

row_id$Var1 <- as.numeric(as.character(row_id$Var1))

important <- test[row_id$Var1, ]

important <- merge(important, taxonomy, by = "ASV_id")

write.csv(important, "./tables/enrichment_families/M-ARCOL_Deschamps_et_al_2020_enriched.csv")
```

# SHIME_Firrman_et_al_2021

```{r}
asv_list_files <- list.files("./tables/enrichment_analysis/SHIME_Firrman_et_al_2021", full.names = TRUE)

asv_sample_id <- file_path_sans_ext(list.files("./tables/enrichment_analysis/SHIME_Firrman_et_al_2021"))

csv <- lapply(asv_list_files, read.csv)

project_table <- as.data.frame(csv[[1]])

for (i in 2:length(csv))
{
  project_table <- full_join(project_table, as.data.frame(csv[[i]]))
}

colnames(project_table) <- c("ASV_id", asv_sample_id)

# convert to presence/absence data.frame (i.e. >)

project_table[is.na(project_table)] <- 0


project_table[, 2:length(project_table)] <-
  ifelse(project_table[, 2:length(project_table)] > 1, 1, 0)

test <- project_table

test <- test[, -c(2, 5)]

idx <- data.frame(which(test != 0, arr.ind = TRUE))

row_id <- data.frame(table(idx$row))

row_id <- row_id %>% filter(Freq > 2)

row_id$Var1 <- as.numeric(as.character(row_id$Var1))

important <- test[row_id$Var1, ]

important <- merge(important, taxonomy, by = "ASV_id")

write.csv(important, "./tables/enrichment_families/SHIME_Firrman_et_al_2021_enriched.csv")
```

# SHIME_Liu_et_al_2022

```{r}
asv_list_files <- list.files("./tables/enrichment_analysis/SHIME_Liu_et_al_2022", full.names = TRUE)

asv_sample_id <- file_path_sans_ext(list.files("./tables/enrichment_analysis/SHIME_Liu_et_al_2022"))

csv <- lapply(asv_list_files, read.csv)

project_table <- as.data.frame(csv[[1]])

for (i in 2:length(csv))
{
  project_table <- full_join(project_table, as.data.frame(csv[[i]]))
}

colnames(project_table) <- c("ASV_id", asv_sample_id)

# convert to presence/absence data.frame (i.e. >)

project_table[is.na(project_table)] <- 0


project_table[, 2:length(project_table)] <-
  ifelse(project_table[, 2:length(project_table)] > 1, 1, 0)

test <- project_table

test <- test[, -c(2, 5)]

idx <- data.frame(which(test != 0, arr.ind = TRUE))

row_id <- data.frame(table(idx$row))

row_id <- row_id %>% filter(Freq > 2)

row_id$Var1 <- as.numeric(as.character(row_id$Var1))

important <- test[row_id$Var1, ]

important <- merge(important, taxonomy, by = "ASV_id")

write.csv(important, "./tables/enrichment_families/SHIME_Liu_et_al_2022_enriched.csv")
```

# SIMGI_Zorraquin_et_al_2021

```{r}
asv_list_files <- list.files("./tables/enrichment_analysis/SIMGI_Zorraquin_et_al_2021/", full.names = TRUE)

asv_sample_id <- file_path_sans_ext(list.files("./tables/enrichment_analysis/SIMGI_Zorraquin_et_al_2021"))

csv <- lapply(asv_list_files, read.csv)

project_table <- as.data.frame(csv[[1]])

for (i in 2:length(csv))
{
  project_table <- full_join(project_table, as.data.frame(csv[[i]]))
}

colnames(project_table) <- c("ASV_id", asv_sample_id)

# convert to presence/absence data.frame (i.e. >)

project_table[is.na(project_table)] <- 0


project_table[, 2:length(project_table)] <-
  ifelse(project_table[, 2:length(project_table)] > 1, 1, 0)

test <- project_table

test <- test[, -c(2, 5)]

idx <- data.frame(which(test != 0, arr.ind = TRUE))

row_id <- data.frame(table(idx$row))

row_id <- row_id %>% filter(Freq > 2)

row_id$Var1 <- as.numeric(as.character(row_id$Var1))

important <- test[row_id$Var1, ]

important <- merge(important, taxonomy, by = "ASV_id")

write.csv(important, "./tables/enrichment_families/SIMGI_Zorraquin_et_al_2021_enriched.csv")
```
